git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk/gtk+@18619 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ gdk/win32/gdkmain-win32.c
@@ -25,7 +25,7 @@
  * GTK+ at ftp://ftp.gtk.org/pub/gtk/. 
  */
 
-#include <config.h>
+#include "config.h"
 
 #include <glib/gprintf.h>
 #include <stdlib.h>
@@ -368,7 +368,7 @@ _gdk_win32_print_dc (HDC hdc)
 	   _gdk_win32_psstyle_to_string (extlogpen.elpPenStyle),
 	   _gdk_win32_psendcap_to_string (extlogpen.elpPenStyle),
 	   _gdk_win32_psjoin_to_string (extlogpen.elpPenStyle),
-	   extlogpen.elpWidth,
+	   (int) extlogpen.elpWidth,
 	   _gdk_win32_lbstyle_to_string (extlogpen.elpBrushStyle));
   g_print ("rop2: %s textcolor=%06lx\n",
 	   _gdk_win32_rop2_to_string (GetROP2 (hdc)),
